# depot-lifecycle-openapi-2.2.5.yaml
openapi: 3.0.1
info:
  title: Depot Life Cycle
  description: "# Purpose\n\n A depot centric API for managing the interchange activity\
    \ & repair lifecycle of a shipping container.  The API is expected to be used\
    \ by Customers, Depots, and Owners to facilitate real time communication between\
    \ systems instead of traditional EDI files.\n\n\n\n # Overview\n\n The depot lifecycle\
    \ API is a RESTful API.  It's requests & responses are loosely based upon traditional\
    \ EDI files.  For code definitions, explanations, or traditional EDI definitions\
    \ refer to [IICL TB 002, February 2003](https://www.iicl.org/iiclforms/assets/File/public/bulletins/TB002_EDIS_February_2003.pdf).\n\
    \n\n\n # JSON\n\n The requests and responses of this API are formatted according\
    \ to the [JSON](https://www.json.org/) standard.\n\n ## Null vs Absent\n\n For\
    \ all types other than arrays, nullable and absent are equivalent by this specification.\
    \  For example, if a non-array attribute is marked optional it may be excluded\
    \ from the JSON or it may be explicitly set to null.  However, to keep request\
    \ sizes small, it is recommended to prefer excluding attributes rather than sending\
    \ an explicit null.\n\n For arrays, null should be avoided.  For example, if the\
    \ array is empty or there are no values expected for the array then either the\
    \ field can be excluded from the JSON or an empty array should be transmitted.\n\
...
  contact:
...
servers:
...
security:
- Dynamic_Token: []
- Static_Token: []
...
paths:
...
  /api/v2/release:
    get:
      tags:
      - release
      summary: search for a release
      description: Finds Releases for the given the criteria.
      operationId: indexRelease
      parameters:
      - name: releaseNumber
        in: query
        description: the release number to filter to
        schema:
          maxLength: 16
          type: string
          nullable: true
          example: RHAMG000000
      - name: includeCandidates
        in: query
        description: whether to include candidate units for any found release
        schema:
          type: boolean
          nullable: true
          example: false
      - name: gateCheck
        in: query
        description: flag to indicate this search is to check if the found advices
          are valid for gate out
        schema:
          type: boolean
          nullable: true
          example: true
          default: true
      responses:
        "200":
          description: successful search
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Release"
        "400":
          description: an error occurred
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: security disallows access
        "501":
          description: this feature is not supported by this server
        "503":
          description: "API is temporarily paused, and not accepting any activity"
      security:
      - JWT_TOKEN: []
      x-iicl-purpose: reporting
    post:
      tags:
      - release
      summary: create release
      description: Creates a Release for the given criteria.
      operationId: saveRelease
      requestBody:
        description: Data to use to update the given Release
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Release"
        required: true
      responses:
        "200":
          description: successful create
          content:
            application/json:
              schema: {}
        "400":
          description: an error occurred
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: security disallows access
        "404":
          description: the release depot was not found
        "501":
          description: this feature is not supported by this server
        "503":
          description: "API is temporarily paused, and not accepting any activity"
      security:
      - JWT_TOKEN: []
      x-iicl-purpose: activity
  /api/v2/release/{releaseNumber}:
    put:
      tags:
      - release
      summary: update release
      description: Updates an existing Release.
      operationId: updateRelease
      parameters:
      - name: releaseNumber
        in: path
        description: name that need to be updated
        required: true
        schema:
          maxLength: 16
          type: string
          example: RHAMG000000
      requestBody:
        description: Data to use to update the given Release
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Release"
        required: true
      responses:
        "200":
          description: successful update
          content:
            application/json:
              schema: {}
        "400":
          description: an error occurred
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: security disallows access
        "404":
          description: the release was not found
        "501":
          description: this feature is not supported by this server
        "503":
          description: "API is temporarily paused, and not accepting any activity"
      security:
      - JWT_TOKEN: []
      x-iicl-purpose: activity
...
components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        code:
          maxLength: 6
          pattern: "^[A-Z0-9]{3}[0-9]{3}$"
          required:
          - "false"
          type: string
          description: indicator code specific to this error
          nullable: true
          example: TRI521
        message:
          required:
          - "false"
          type: string
          description: a descriptive error message
          nullable: true
          example: Info TRI521 - Unit has been gated-in but is not off-hired
        details:
          required:
          - "false"
          type: array
          description: details for the descriptive error message
          nullable: false
          items:
            type: string
            example: "Customer may turn in unit, but will continued to be billed."
      description: |-
        a general response object used when an error occurs to describe why it occurred

        `400` http status codes are used when a request is invalid or the basic requirements are not met

        `405` http status codes are used for business validations
...
    Party:
      required:
      - companyId
      type: object
      properties:
        companyId:
          maxLength: 9
          pattern: "^[A-Z0-9]{9}$"
          required:
          - "true"
          type: string
          description: "the identifier for this party, often referred to as an EDI\
            \ Address"
          nullable: false
          example: DEHAMCMRA
        userCode:
          maxLength: 16
          required:
          - "false"
          type: string
          description: the user identifier at this company that concerns this message
          nullable: true
          example: JDOE
        userName:
          maxLength: 70
          required:
          - "false"
          type: string
          description: the full name for the user identified by `userCode`
          nullable: true
          example: John Doe
        faxNumber:
          required:
          - "false"
          type: array
          description: the contact fax number(s) for this party
          nullable: false
          items:
            type: string
        phoneNumber:
          required:
          - "false"
          type: array
          description: the contact phone number(s) for this party
          nullable: false
          items:
            type: string
        emailAddress:
          required:
          - "false"
          type: array
          description: the contact email address(es) for this party
          nullable: false
          items:
            type: string
        name:
          maxLength: 150
          required:
          - "false"
          type: string
          description: the name of this company
          nullable: true
          example: CMR Container Maintenance Rep.
        code:
          maxLength: 10
          required:
          - "false"
          type: string
          description: "the internal system code for this company, will be system\
            \ specific to the system delivering or receiving this message"
          nullable: true
          example: HAMG
        streetAddress:
          required:
          - "false"
          type: array
          description: the street address lines
          nullable: false
          items:
            type: string
        city:
          maxLength: 28
          required:
          - "false"
          type: string
          description: the city for the address
          nullable: true
        country:
          maxLength: 2
          required:
          - "false"
          type: string
          description: the 2 digit ISO country code for the address
          nullable: true
        postalCode:
          maxLength: 20
          required:
          - "false"
          type: string
          description: the postal code for the address
          nullable: true
        stateProvince:
          maxLength: 20
          required:
          - "false"
          type: string
          description: the optional state or province code for the address
          nullable: true
        latitude:
          required:
          - "false"
          type: number
          description: "Instead of an address, provides the latitude of this party"
          format: double
          nullable: true
        longitude:
          required:
          - "false"
          type: number
          description: "Instead of an address, provides the longitude of this party"
          format: double
          nullable: true
      description: represents a company (or location) involving shipping containers
...
    Release:
      required:
      - approvalDate
      - depot
      - details
      - expirationDate
      - owner
      - quantity
      - recipient
      - releaseNumber
      - status
      - type
      type: object
      properties:
        releaseNumber:
          maxLength: 16
          required:
          - "true"
          type: string
          description: "the identifier for this release, also referred to as the advice\
            \ number or release number"
          nullable: false
          example: AHAMG33141
        status:
          description: "Describes the status for this release: \n\n`PENDING` - release\
            \ is pending and not yet valid\n\n`APPROVED` - release is approved for\
            \ leaving the storage location\n\n`COMPLETE` - all containers have left\
            \ the depot and no more may be processed\n\n`EXPIRED` - the release is\
            \ now expired and any remaining units are no longer valid\n\n`CANCELLED`\
            \ - the release is cancelled and not valid"
          example: APPROVED
          allOf:
          - $ref: "#/components/schemas/ReleaseStatus"
        type:
          description: "Describes the intended purpose of the release: \n\n`SALE`\
            \ - shipping containers are being sold\n\n`BOOK` - shipping containers\
            \ are being leased to a customer\n\n`REPO` - shipping containers are being\
            \ relocated by the owner to another storage location"
          allOf:
          - $ref: "#/components/schemas/ReleaseType"
        approvalDate:
          required:
          - "true"
          type: string
          description: |-
            the date and time in the depot local time zone (i.e. `2019-07-21T17:32:28Z`) that this release is considered approved / effective

            ( notation as defined by [RFC 3339, section 5.6](https://tools.ietf.org/html/rfc3339#section-5.6) )
          format: date-time
          nullable: false
        expirationDate:
          required:
          - "true"
          type: string
          description: |-
            the date and time in the depot local time zone (i.e. `2020-07-21T17:32:28Z`) that this release is considered no longer valid

            ( notation as defined by [RFC 3339, section 5.6](https://tools.ietf.org/html/rfc3339#section-5.6) )
          format: date-time
          nullable: false
        comments:
          required:
          - "false"
          type: array
          description: comments pertaining to this unit for the intended recipient
            of this message
          nullable: false
          items:
            type: string
            example: An example release level comment.
        depot:
          description: The location for this release
          allOf:
          - $ref: "#/components/schemas/Party"
        owner:
          description: The owner of the shipping container that approved the release
          allOf:
          - $ref: "#/components/schemas/Party"
        recipient:
          description: The intended recipient for this message representing a release
          allOf:
          - $ref: "#/components/schemas/Party"
        onHireSurveyRequired:
          type: boolean
          description: indicates if an on-hire survey is required for units associated
            to this release
        quantity:
          minimum: 0
          required:
          - "true"
          type: integer
          description: the number of shipping containers assigned to this release.
          format: int32
          example: 1
        details:
          minItems: 1
          required:
          - "true"
          type: array
          description: groups of like-criteria units
          nullable: false
          items:
            $ref: "#/components/schemas/ReleaseDetail"
      description: An approval to remove shipping containers from a storage location.
    ReleaseDetail:
      required:
      - contract
      - customer
      - equipment
      - grade
      - quantity
      type: object
      properties:
        customer:
          description: The customer for the contract on this detail.
          allOf:
          - $ref: "#/components/schemas/Party"
        contract:
          maxLength: 16
          required:
          - "true"
          type: string
          description: the contract code for the given shipping containers
          nullable: false
          example: CNCX05-100000
        equipment:
          maxLength: 10
          required:
          - "true"
          type: string
          description: the equipment type ISO code or an internal code if one does
            not exist for the given shipping containers
          nullable: false
          example: 22G1
        grade:
          maxLength: 10
          required:
          - "true"
          type: string
          description: the current grade of the unit
          nullable: false
          example: IICL
        upgradeType:
          description: |-
            the type of secondary upgrade this estimate represents.

            `FG` - Food grade

            `ML` - Malt

            `DB` - Dairy Board

            `EV` - Evian

            `WH` - Whiskey

            `SU` - Sugar

            `CF` - Coffee

            `TB` - Tobacco

            `MC` - Milk cartons

            `MP` - Milk powder

            `AM` - Ammunition

            `CH` - Cotton/Hay

            `TE` - Tea

            `FT` - Flexitank
          nullable: true
          example: AM
          allOf:
          - $ref: "#/components/schemas/UpgradeType"
        units:
          required:
          - "false"
          type: array
          description: "the specific units for this release if defined, if not, assumed\
            \ blanket (any unit matching criteria can be tied up to the quantity limit\
            \ of this detail)"
          nullable: false
          items:
            $ref: "#/components/schemas/ReleaseUnit"
        criteria:
          required:
          - "false"
          type: array
          description: 'additional criteria beyond the required properties of this
            detail to further restrict units.  i.e. <= 2003 manufacture year. '
          nullable: false
          items:
            $ref: "#/components/schemas/ReleaseDetailCriteria"
        comments:
          required:
          - "false"
          type: array
          description: comments pertaining to this unit for the intended recipient
            of this message
          nullable: false
          items:
            type: string
            example: An example detail level comment.
        quantity:
          minimum: 0
          required:
          - "true"
          type: integer
          description: the number of shipping containers assigned to this detail
          format: int32
          nullable: false
          example: 1
        preTripInspectionRequired:
          required:
          - "false"
          type: boolean
          description: indicator if mechanical equipment must be tested prior to lease
            out
          nullable: true
        desiredTemperature:
          required:
          - "false"
          type: integer
          description: the reefer setpoint / desired temperature in Celsius
          format: int32
          nullable: true
          example: -23
        ventilation:
          maxLength: 10
          required:
          - "false"
          type: string
          description: "if the equipment has fresh air ventilation, the rate of the\
            \ fresh air ventilation"
          nullable: true
          example: 90 CBM
      description: groups similar units on a release
    ReleaseDetailCriteria:
      required:
      - fieldDescription
      - fieldId
      - fieldName
      - fieldValue
      type: object
      properties:
        fieldName:
          maxLength: 50
          required:
          - "true"
          type: string
          description: a legible field name
          nullable: false
          example: Minimum Tonnage
        fieldId:
          maxLength: 30
          required:
          - "true"
          type: string
          description: an abbreviated form of the field name
          nullable: false
          example: tonnageMin
        fieldDescription:
          maxLength: 255
          required:
          - "true"
          type: string
          description: an explanation of what the field value is
          nullable: false
          example: minimum tonnage rating of the unit (inclusive)
        fieldValue:
          maxLength: 500
          required:
          - "true"
          type: string
          description: the value of the field
          nullable: false
          example: "30"
      description: represents unit criteria to restrict units on a given detail
    ReleaseStatus:
      type: string
      description: "Describes the status for this release: \n\n`PENDING` - release\
        \ is pending and not yet valid\n\n`APPROVED` - release is approved for leaving\
        \ the storage location\n\n`COMPLETE` - all containers have left the depot\
        \ and no more may be processed\n\n`EXPIRED` - the release is now expired and\
        \ any remaining units are no longer valid\n\n`CANCELLED` - the release is\
        \ cancelled and not valid"
      example: APPROVED
      enum:
      - PENDING
      - APPROVED
      - COMPLETE
      - EXPIRED
      - CANCELLED
    ReleaseType:
      type: string
      description: "Describes the intended purpose of the release: \n\n`SALE` - shipping\
        \ containers are being sold\n\n`BOOK` - shipping containers are being leased\
        \ to a customer\n\n`REPO` - shipping containers are being relocated by the\
        \ owner to another storage location"
      enum:
      - SALE
      - BOOK
      - REPO
    ReleaseUnit:
      required:
      - status
      - unitNumber
      type: object
      properties:
        unitNumber:
          maxLength: 11
          pattern: "^[A-Z]{4}[X0-9]{6}[A-Z0-9]{0,1}$"
          required:
          - "true"
          type: string
          description: the current unit number of the shipping container
          nullable: false
          example: CONU1234561
        comments:
          required:
          - "false"
          type: array
          description: comments pertaining to this unit for the intended recipient
            of this message
          nullable: false
          items:
            type: string
            example: An example unit level comment.
        status:
          description: "Describes the state of the shipping container for this release:\
            \ \n\n`TIED` - shipping container is assigned to this release and ready\
            \ to lease out.\n\n`REMOVED` - shipping container was attached to this\
            \ release, but is no longer valid for release.\n\n`LOT` - shipping container\
            \ has left the storage location.\n\n`CANDIDATE` - this container is eligible\
            \ for this release but not currently assigned."
          example: TIED
          allOf:
          - $ref: "#/components/schemas/ReleaseUnitStatus"
        carbCompliant:
          required:
          - "false"
          type: boolean
          description: "specific to Genset equipment, indicator for California Air\
            \ Resources Board compliance"
          nullable: true
        manufactureDate:
          required:
          - "false"
          type: string
          description: |-
            date and month this unit was manufactured

            ( full-date notation as defined by [RFC 3339, section 5.6](https://tools.ietf.org/html/rfc3339#section-5.6) )
          format: date
          nullable: true
          example: 2001-07-21
      description: information for a specific unit on a release
    ReleaseUnitStatus:
      type: string
      description: "Describes the state of the shipping container for this release:\
        \ \n\n`TIED` - shipping container is assigned to this release and ready to\
        \ lease out.\n\n`REMOVED` - shipping container was attached to this release,\
        \ but is no longer valid for release.\n\n`LOT` - shipping container has left\
        \ the storage location.\n\n`CANDIDATE` - this container is eligible for this\
        \ release but not currently assigned."
      example: TIED
      enum:
      - REMOVED
      - TIED
      - LOT
      - CANDIDATE
...
    UpgradeType:
      type: string
      description: |-
        an indicator for the upgrades applied to units on this detail.

        `FG` - Food grade

        `ML` - Malt

        `DB` - Dairy Board

        `EV` - Evian

        `WH` - Whiskey

        `SU` - Sugar

        `CF` - Coffee

        `TB` - Tobacco

        `MC` - Milk cartons

        `MP` - Milk powder

        `AM` - Ammunition

        `CH` - Cotton/Hay

        `TE` - Tea

        `FT` - Flexitank
      example: AM
      enum:
      - FG
      - ML
      - DB
      - EV
      - WH
      - SU
      - CF
      - TB
      - MC
      - MP
      - AM
      - CH
      - TE
      - FT
...
  securitySchemes:
    JWT_TOKEN:
      type: http
      description: JWT Bearer Authentication
      scheme: bearer
      bearerFormat: JWT
...